kickout 3 tests from e2e_frequency
